outsource minimal_model to separate file
